+++ NEWS.md
@@ -1,5 +1,7 @@
 ## 2.0.0 (IN PROGRESS)
--
+- Derive MARC 300 from Graph [MODLD-401](https://folio-org.atlassian.net/browse/MODLD-401)
+- Update Mapping for 300$b & $e (Physical description & Accompanying material) [MODLD-508](https://folio-org.atlassian.net/browse/MODLD-508)
+- Convert marc field 586 (Awards Note) to Resource description [MODLD-366](https://folio-org.atlassian.net/browse/MODLD-366)
 
 ## 1.0.2 (03-12-2025)
 - Initial release